Mudança de JDBC para JPA

+++ VilaAcate/src/main/java/com/devinhouse/vilacate/rest/PeopleRest.java
@@ -5,6 +5,7 @@ import com.devinhouse.vilacate.model.dto.PeopleDto;
 import com.devinhouse.vilacate.model.entity.People;
 import com.devinhouse.vilacate.services.FinancialService;
 import com.devinhouse.vilacate.services.PeopleService;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,13 +17,18 @@ import java.util.List;
 @RequestMapping("vilacate")
 public class PeopleRest {
 
-    @Autowired
-    PeopleService peopleService;
-    @Autowired
-    FinancialService financialService;
+    private PeopleService peopleService;
+    private FinancialService financialService;
+    private RabbitTemplate rabbitTemplate;
+
+    public PeopleRest(PeopleService peopleService, FinancialService financialService, RabbitTemplate rabbitTemplate) {
+        this.peopleService = peopleService;
+        this.financialService = financialService;
+        this.rabbitTemplate = rabbitTemplate;
+    }
 
     @GetMapping("/list")
-    public List<People> listAllPeoples(@RequestParam(name = "mes", required = false) Integer month) {
+    public List<PeopleDto> listAllPeoples(@RequestParam(name = "mes", required = false) Integer month) {
         if (month != null) {
             return this.peopleService.listByMonth(month);
         }
@@ -35,7 +41,7 @@ public class PeopleRest {
     }
 
     @GetMapping("/list/name")
-    public List<People> listByName(@RequestParam(name = "nome") String name){
+    public List<PeopleDto> listByName(@RequestParam(name = "nome") String name){
         return this.peopleService.peopleByName(name);
     }
 
@@ -48,14 +54,14 @@ public class PeopleRest {
     @GetMapping("/rent")
     public FinancialDto showFinancial(){
         FinancialDto financialDto = this.financialService.showFinancial();
-        System.out.println(financialDto.getVillageCost());
+        rabbitTemplate.convertAndSend("FinancialExchange","acate.*", financialDto);
         return financialDto;
     }
 
     @PostMapping("/add")
     public ResponseEntity<HttpStatus> createPeople(@RequestBody PeopleDto peopleDto){
-        Boolean created = this.peopleService.createPeople(peopleDto);
-        if(created){
+        People created = this.peopleService.createPeople(peopleDto);
+        if(created != null){
             return ResponseEntity.status(HttpStatus.CREATED).build();
         }
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
@@ -63,11 +69,13 @@ public class PeopleRest {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<HttpStatus> deletePeople(@PathVariable("id") Long id) {
-        Boolean deleted = this.peopleService.deletePeople(id);
-        if (deleted){
-            return ResponseEntity.status(HttpStatus.OK).build();
+        if (id == null){
+            throw new IllegalArgumentException("Erro ID vazio!");
         }
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+
+        this.peopleService.deletePeople(id);
+
+        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
 }
